code/back-end/django_server/requirements.txt
appnope==0.1.2
argon2-cffi==21.1.0
asgiref==3.5.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
boto3==1.23.4
botocore==1.26.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.11.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-annoying==0.10.6
django-cors-headers==3.11.0
django-heroku==0.3.1
django-storages==1.12.3
djangorestframework==3.13.1
entrypoints==0.3
filelock==3.6.0
fonttools==4.28.5
gunicorn==20.1.0
idna==3.2
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jmespath==1.0.0
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.26
mysqlclient==2.1.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.5.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
psycopg2==2.9.3
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pymssql==2.2.2
pyodbc==4.0.32
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.7.3
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
virtualenv==20.13.2
wcwidth==0.2.5
webdriver-manager==3.5.0
webencodings==0.5.1
whitenoise==6.0.0
widgetsnbextension==3.5.1
